django-calendar/requirements.txt
asgiref==3.6.0
Django==4.2
django-admin-autocomplete-filter==0.7.1
django-admin-rangefilter==0.11.0
django-colorfield==0.9.0
Pillow==10.0.0
sqlparse==0.4.3
tzdata==2023.3
